requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
qrcode[pil]==7.4.2
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0

Flask==2.1.1
Flask-Cors==3.1.1